+++ CHANGELOG.md
@@ -7,7 +7,13 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Added
 
+* RedirectResponse which implements ResponseInterface
 * Available values now can have description in Console and OpenApiHandler. Just use associative array - keys are available values, values are their description
+
+### Fixed
+
+* Yaml format for OpenApiHandler is available only if symfony/yaml is installed
+* RedirectOutput use new RedirectResponse
 
 ## 2.6.0 - 2021-10-08