+++ CHANGELOG.md
@@ -4,9 +4,11 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased][unreleased]
 
+
 ### Added
 
 * RedirectResponse which implements ResponseInterface
+* Available values now can have description in Console and OpenApiHandler. Just use associative array - keys are available values, values are their description
 
 ### Fixed