+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased][unreleased]
 
+#### Added
+
+* PHP 8.1 support (`league/fractal` 0.21.2) (BC Break)
+
 #### Fixed
 
 * Missing expiration params in XmlApiResponse