+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 #### Added
 
-* PHP 8.1 support (`league/fractal` 0.20.1) (BC Break)
+* PHP 8.1 support (`league/fractal` 0.20.1)
 
 #### Fixed